refactor #97: 응답코드 및 설정 파일 패키지 위치 수정
