+++ user-service/src/main/java/com/musinsam/userservice/app/infrastructure/config/PasswordEncoderConfig.java
@@ -1,10 +1,9 @@
-package com.musinsam.userservice.app.global.config;
+package com.musinsam.userservice.app.infrastructure.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Component;
 
 @Configuration
 public class PasswordEncoderConfig {
